WIP - splitting up code for better maintainability
